requirements.txt
asgiref==3.2.3
autopep8==1.4.4
blessings==1.7
bpython==0.18
certifi==2019.11.28
chardet==3.0.4
curtsies==0.3.0
Django==3.0
django-extensions==2.2.5
django-resized==0.3.9
djangorestframework==3.11.0
greenlet==0.4.15
idna==2.8
Pillow==7.1.1
pycodestyle==2.5.0
pydotplus==2.0.2
Pygments==2.5.2
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
typing==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.7
